Fixed tests and code cleanup
