+++ src/components/StopwatchButton/StopwatchButton.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { StyleSheet, View, TouchableOpacity, Text } from 'react-native';
-import { StopwatchButtonProps } from './types/types';
+import { View, TouchableOpacity, Text } from 'react-native';
+import { StopwatchButtonProps } from '../../types/types';
+import styles from './styles';
 
 export default function StopwatchButton({ isRunning, hasStarted, onStart, onStop, onPause, onReset, onLap }: StopwatchButtonProps) {
 
@@ -38,28 +39,3 @@ export default function StopwatchButton({ isRunning, hasStarted, onStart, onStop
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    marginTop: 20,
-  },
-  button: {
-    width: 80, // Fixed width for a circular shape
-    height: 80, // Fixed height for a circular shape
-    borderRadius: 40, // Half of width/height to make it circular
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#333',
-    marginHorizontal: 10,
-  },
-  buttonText: {
-    color: '#FFF',
-    fontSize: 16,
-  },
-  stopButton: {
-    backgroundColor: '#D00',
-  }
-});
-
